+++ tools/netcat-client/admin/create-request-reload-config.php
@@ -4,8 +4,8 @@ $serverKey = file_get_contents("serverkey");
 $request = json_encode(array(
     "_t" => round(microtime(true)),
     "type" => 1, // RELOAD_CONFIG
+    "logger_id" => "sample-app",
 ));
 
 echo shell_exec("echo '$request' | ripe -e --key $serverKey");
-// Following will fail as we do not allow admin requests using server public key
-//echo shell_exec("echo '$request' | ripe -e --in-key server-1024-public.pem") . "\r\n\r\n";
+//printf("echo '$request' | ripe -e --key $serverKey");